Update chapter_1.md
Okidoki - I love this concept so much. But my slang is outdated and parts I didn't know how to change are direct from the book. This is the entirety of the first chapter. My editing style has always been to take the existing text and change it as I can. I dunno if you'll like this or not, so here's hoping.